fix call to /null url
